Исправление ошибок, перенос констант, уход от async/await к .then
